Shopper-python/shopapp/requirements.txt
playwright
pydantic
nest_asyncio
python-dotenv
langchain
langchain-openai
fastapi
uvicorn
